feat(APIUtil): Add FancyPluginBase
Add FancyPluginBase which can be extended from to register Listeners on
    attributes of Items
Converted void Init() to Task Init()

The Problem that exists with PluginBase is, that it is stateless. If one
want's to compare different states / changes of a User this has not been
possible previously.
With FancyPluginBase one can register a listener together with a mapper.
The mapper receives the object registered on during every change and
can map that object to a subset of it's fields. Only when this subset
(i.e. at least one of the mapped fields) changes, the listener gets
called with the new (modified) object.